Consolidate <LinkText> component
