+++ src/lib/Components/Bidding/Components/LinkText.tsx
@@ -1,0 +1,5 @@
+import styled from "styled-components/native"
+
+export const LinkText = styled.Text`
+  text-decoration-line: underline;
+`
